+++ CHANGELOG.md
@@ -6,6 +6,7 @@ There are some key used within the changelog which are as following:
 
 # IN PROGRESS
 
+- fixed issue with allow create not working for extend texts given a static set of options that are empty #168
 - fixed issue with clicking popover handle trigger the on click outside event of the popover itself
 - fixed issue with popover and weird rendering issue when mouseenter / mouseleave event (like a tooltip) #167
 - refactored demo application side navigation menu to use new expandable list component